refactor: address feedback, part 1
