+++ src/common/log-helpers.ts
@@ -1,5 +1,4 @@
 import { join } from 'path';
-import readline, { ReadLine } from 'readline';
 import { FileLog } from './file-log';
 
 export function getDefaultLogPath(type: string, action: string, platform: string = process.platform): string {
@@ -20,26 +19,4 @@ export function createLog(logFile: string, title?: string): FileLog {
   }
 
   return log;
-}
-
-function asyncQuestionInternal(rl: ReadLine, question: string): Promise<string> {
-  return new Promise((resolve): void => {
-    rl.question(question, resolve);
-  });
-}
-
-export async function asyncQuestion(question: string, log?: FileLog): Promise<boolean> {
-  const rl = readline.createInterface({
-    input: process.stdin,
-    output: process.stdout,
-    terminal: false
-  });
-
-  const answer = await asyncQuestionInternal(rl, question);
-  rl.close();
-
-  if (log != null) {
-    log.appendLine(question + answer, true);
-  }
-  return answer.length > 0 && answer[0].toLowerCase() === 'y';
 }
